requirements.txt
# package
google-cloud-bigquery
pandas
pandas_gbq
